Split up huge settings module
- Move validators into their own module.
- Convert Kind to IntEnum

Related #2273
